modified shipping cost

+++ models/staging/raw/stg_raw__ship.sql
@@ -12,7 +12,7 @@ renamed as (
         orders_id,
         shipping_fee,
         logcost,
-        cast(ship_cost as integer)
+        cast(ship_cost as float64)
 
     from source
